requirements.txt
discord.py==1.7.3
colorama==0.4.6
aiohttp==3.8.5
async-timeout==4.0.3
multidict==6.0.4
yarl==1.9.2
